requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch==2.6.0+cu126
torchvision==0.21.0+cu126
torchaudio==2.6.0+cu126
fastapi
uvicorn
streamlit
pymupdf
python-docx
beautifulsoup4
sentence-transformers
faiss-cpu
requests==2.32.5
urllib3==2.2.3
charset-normalizer==3.3.2
chardet==5.2.0
pytest
langgraph==1.0.9
langchain-core==1.2.17
pydantic
python-multipart
langchain-text-splitters
qdrant-client
aiohttp
tiktoken
tenacity
python-dotenv
playwright
pillow
numpy
pandas>=2.0.0
openpyxl>=3.1.2
markdownify
tabulate
selectolax
transformers
coqui-tts
easyocr
celery
redis
accelerate
selectolax
pdfplumber>=0.11.0
prometheus-client>=0.20.0
psutil
ragas
paddlepaddle-gpu
langchain-experimental==0.4.1
prophet
scikit-learn
langchain-groq==1.1.2
# Third-Party API Integrations
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
simple-salesforce
stripe
snowflake-connector-python
hubspot-api-client
msal
numpy>=2.0.0
xgboost>=2.0.0
xlsxwriter
scipy==1.11.4
statsmodels
dowhy
econml
river

# Langfuse
langfuse>=2.0.0
langfuse-langchain
langchain-google-genai

# Auth / JWT
python-jose[cryptography]>=3.3.0
